SPE UDOM 2026/backend/requirements.txt
django==6.0.3
channels==4.2.2
channels-redis==4.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
psycopg[binary]==3.2.10
whitenoise==6.9.0
django-ratelimit==4.1.0
cryptography==43.0.0

